requirements.txt
flask
flask_bootstrap
flask-sqlalchemy
flask-wtf
wtforms
flask-bcrypt
flask-login
PyJWT
Flask-Mail